refactor: extract modals into components
